Fixed final restart notification and config.

+++ src/main/java/com/breakinblocks/bbserver/Config.java
@@ -24,7 +24,7 @@ public class Config {
 
         // 2 hours, 1 hour, 30 minutes, 10 minutes, 5 minutes, 1 minutes, 30 seconds, 10 seconds, 5, 4, 3, 2, 1
         @Comment("Notification times in seconds")
-        public static long[] notifications = new long[]{2 * 60 * 60, 60 * 60, 30 * 60, 10 * 60, 5 * 60, 60, 30, 10, 5, 4, 3, 2, 1};
+        public static int[] notifications = new int[]{2 * 60 * 60, 60 * 60, 30 * 60, 10 * 60, 5 * 60, 60, 30, 10, 5, 4, 3, 2, 1};
 
         @Comment("Restart flag file. File is written then the server is shutdown. Server run script is responsible for deleting the file and re-running the server instead of exiting.")
         public static String flag = "autostart.stamp";
